requirements.txt
streamlit
langchain
openai
pypdf
faiss-cpu
python-dotenv
